backend/requirements.txt
django
PyJWT
django-cors-headers
pillow
gunicorn
asgiref